+++ src/be_iolib.c
@@ -7,7 +7,7 @@ static int m_input(bvm *vm)
 {
     size_t pos = 0, size = READLINE_STEP;
     char *buffer = be_malloc(size);
-    char *res = be_fgets(stdin, buffer, size);
+    char *res = be_fgets(stdin, buffer, (int)size);
     while (res) {
         pos += strlen(buffer + pos) - 1;
         if (!pos || buffer[pos] == '\n') {
